typed_path

Enum Utf8WindowsComponent

pub enum Utf8WindowsComponent<'a> {
    Prefix(Utf8WindowsPrefixComponent<'a>),
    RootDir,
    CurDir,
    ParentDir,
    Normal(&'a str),
}

str slice version of std::path::Component that represents a Windows-specific component

Variants

Prefix(Utf8WindowsPrefixComponent<'a>)

RootDir

CurDir

ParentDir

Normal(&'a str)

Implementations

impl<'a> Utf8WindowsComponent<'a>

pub fn as_path<T>(&self) -> &Utf8Path<T>
where T: Utf8Encoding,

Returns path representing this specific component

pub fn from_utf8(component: &WindowsComponent<'a>) -> Result<Self, Utf8Error>

Converts a non-UTF-8 WindowsComponent to a UTF-8 Utf8WindowsComponent by checking that the component contains valid UTF-8.

Errors

Returns Err if the component is not UTF-8 with a description as to why the provided component is not UTF-8.

Examples

Basic usage:

use typed_path::{Utf8Component, WindowsComponent, Utf8WindowsComponent};

// some bytes, in a vector
let component = WindowsComponent::Normal(&[240, 159, 146, 150]);

// We know these bytes are valid, so just use `unwrap()`.
let utf8_component = Utf8WindowsComponent::from_utf8(&component).unwrap();

assert_eq!("💖", utf8_component.as_str());

Incorrect bytes:

use typed_path::{WindowsComponent, Utf8WindowsComponent};

// some invalid bytes, in a vector
let component = WindowsComponent::Normal(&[0, 159, 146, 150]);

assert!(Utf8WindowsComponent::from_utf8(&component).is_err());

See the docs for Utf8Error for more details on the kinds of errors that can be returned.

pub unsafe fn from_utf8_unchecked(component: &WindowsComponent<'a>) -> Self

Converts a non-UTF-8 WindowsComponent to a UTF-8 Utf8WindowsComponent without checking that the string contains valid UTF-8.

See the safe version, from_utf8, for more information.

Safety

The bytes passed in must be valid UTF-8.

Examples

Basic usage:

use typed_path::{Utf8Component, WindowsComponent, Utf8WindowsComponent};

// some bytes, in a vector
let component = WindowsComponent::Normal(&[240, 159, 146, 150]);

let utf8_component = unsafe {
    Utf8WindowsComponent::from_utf8_unchecked(&component)
};

assert_eq!("💖", utf8_component.as_str());

pub fn is_prefix(&self) -> bool

Returns true if represents a prefix

pub fn prefix(self) -> Option<Utf8WindowsPrefixComponent<'a>>

Converts from Utf8WindowsComponent to Option<Utf8WindowsPrefixComponent>

Converts self into an Option<Utf8WindowsPrefixComponent>, consuming self, and discarding if not a Utf8WindowsPrefixComponent

pub fn prefix_kind(self) -> Option<Utf8WindowsPrefix<'a>>

Converts from Utf8WindowsComponent to Option<Utf8WindowsPrefix>

Converts self into an Option<Utf8WindowsPrefix>, consuming self, and discarding if not a Utf8WindowsPrefixComponent whose kind method we invoke