typed_path

Struct PathBuf

pub struct PathBuf<T>
where
    T: Encoding,
{ /* private fields */ }

An owned, mutable path that mirrors std::path::PathBuf, but operatings using an Encoding to determine how to parse the underlying bytes.

This type provides methods like push and set_extension that mutate the path in place. It also implements Deref to Path, meaning that all methods on Path slices are available on PathBuf values as well.

Examples

You can use push to build up a PathBuf from components:

use typed_path::{PathBuf, WindowsEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let mut path = PathBuf::<WindowsEncoding>::new();

path.push(r"C:\");
path.push("windows");
path.push("system32");

path.set_extension("dll");

However, push is best used for dynamic situations. This is a better way to do this when you know all of the components ahead of time:

use typed_path::{PathBuf, WindowsEncoding};

let path: PathBuf<WindowsEncoding> = [r"C:\", "windows", "system32.dll"].iter().collect();

We can still do better than this! Since these are all strings, we can use From::from:

use typed_path::{PathBuf, WindowsEncoding};

let path = PathBuf::<WindowsEncoding>::from(br"C:\windows\system32.dll");

Which method works best depends on what kind of situation you’re in.

Implementations

impl<T> PathBuf<T>
where T: Encoding,

pub fn new() -> Self

Allocates an empty PathBuf.

Examples

use typed_path::{PathBuf, UnixEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let path = PathBuf::<UnixEncoding>::new();

pub fn with_capacity(capacity: usize) -> Self

Creates a new PathBuf with a given capacity used to create the internal Vec<u8>. See with_capacity defined on Vec.

Examples

use typed_path::{PathBuf, UnixEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let mut path = PathBuf::<UnixEncoding>::with_capacity(10);
let capacity = path.capacity();

// This push is done without reallocating
path.push(r"C:\");

assert_eq!(capacity, path.capacity());

pub fn as_path(&self) -> &Path<T>

Coerces to a Path slice.

Examples

use typed_path::{Path, PathBuf, UnixEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let p = PathBuf::<UnixEncoding>::from("/test");
assert_eq!(Path::new("/test"), p.as_path());

pub fn push<P: AsRef<Path<T>>>(&mut self, path: P)

Extends self with path.

If path is absolute, it replaces the current path.

With WindowsPathBuf:

if path has a root but no prefix (e.g., \windows), it replaces everything except for the prefix (if any) of self.

if path has a prefix but no root, it replaces self.

if self has a verbatim prefix (e.g. \\?\C:\windows) and path is not empty, the new path is normalized: all references to . and .. are removed.

Examples

Pushing a relative path extends the existing path:

use typed_path::{PathBuf, UnixEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let mut path = PathBuf::<UnixEncoding>::from("/tmp");
path.push("file.bk");
assert_eq!(path, PathBuf::from("/tmp/file.bk"));

Pushing an absolute path replaces the existing path:

use typed_path::{PathBuf, UnixEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let mut path = PathBuf::<UnixEncoding>::from("/tmp");
path.push("/etc");
assert_eq!(path, PathBuf::from("/etc"));

pub fn push_checked<P: AsRef<Path<T>>>( &mut self, path: P, ) -> Result<(), CheckedPathError>

Like PathBuf::push, extends self with path, but also checks to ensure that path abides by a set of rules.

Rules

path cannot contain a prefix component.

path cannot contain a root component.

path cannot contain invalid filename bytes.

path cannot contain parent components such that the current path would be escaped.

Examples

Pushing a relative path extends the existing path:

use typed_path::{PathBuf, UnixEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let mut path = PathBuf::<UnixEncoding>::from("/tmp");

// Pushing a relative path works like normal
assert!(path.push_checked("file.bk").is_ok());
assert_eq!(path, PathBuf::from("/tmp/file.bk"));

Pushing a relative path that contains unresolved parent directory references fails with an error:

use typed_path::{CheckedPathError, PathBuf, UnixEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let mut path = PathBuf::<UnixEncoding>::from("/tmp");

// Pushing a relative path that contains parent directory references that cannot be
// resolved within the path is considered an error as this is considered a path
// traversal attack!
assert_eq!(path.push_checked(".."), Err(CheckedPathError::PathTraversalAttack));
assert_eq!(path, PathBuf::from("/tmp"));

Pushing an absolute path fails with an error:

use typed_path::{CheckedPathError, PathBuf, UnixEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let mut path = PathBuf::<UnixEncoding>::from("/tmp");

// Pushing an absolute path will fail with an error
assert_eq!(path.push_checked("/etc"), Err(CheckedPathError::UnexpectedRoot));
assert_eq!(path, PathBuf::from("/tmp"));

pub fn pop(&mut self) -> bool

Truncates self to self.parent.

Returns false and does nothing if self.parent is None. Otherwise, returns true.

Examples

use typed_path::{Path, PathBuf, UnixEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let mut p = PathBuf::<UnixEncoding>::from("/spirited/away.rs");

p.pop();
assert_eq!(Path::new("/spirited"), p);
p.pop();
assert_eq!(Path::new("/"), p);

pub fn set_file_name<S: AsRef<[u8]>>(&mut self, file_name: S)

Updates self.file_name to file_name.

If self.file_name was None, this is equivalent to pushing file_name.

Otherwise it is equivalent to calling pop and then pushing file_name. The new path will be a sibling of the original path. (That is, it will have the same parent.)

Examples

use typed_path::{PathBuf, UnixEncoding};

// NOTE: A pathbuf cannot be created on its own without a defined encoding
let mut buf = PathBuf::<UnixEncoding>::from("/");
assert!(buf.file_name() == None);
buf.set_file_name("bar");
assert!(buf == PathBuf::from("/bar"));
assert!(buf.file_name().is_some());
buf.set_file_name("baz.txt");
assert!(buf == PathBuf::from("/baz.txt"));

pub fn set_extension<S: AsRef<[u8]>>(&mut self, extension: S) -> bool

Updates self.extension to extension.

Returns false and does nothing if self.file_name is None, returns true and updates the extension otherwise.

If self.extension is None, the extension is added; otherwise it is replaced.

Examples

use typed_path::{Path, PathBuf, UnixEncoding};

let mut p = PathBuf::<UnixEncoding>::from("/feel/the");

p.set_extension("force");
assert_eq!(Path::new("/feel/the.force"), p.as_path());

p.set_extension("dark_side");
assert_eq!(Path::new("/feel/the.dark_side"), p.as_path());

pub fn into_vec(self) -> Vec<u8>

Consumes the PathBuf, yielding its internal Vec<u8> storage.

Examples

use typed_path::{PathBuf, UnixEncoding};

let p = PathBuf::<UnixEncoding>::from("/the/head");
let vec = p.into_vec();

pub fn into_boxed_path(self) -> Box<Path<T>>

Converts this PathBuf into a boxed Path.

pub fn capacity(&self) -> usize

Invokes capacity on the underlying instance of Vec.

pub fn clear(&mut self)

Invokes clear on the underlying instance of Vec.

pub fn reserve(&mut self, additional: usize)

Invokes reserve on the underlying instance of Vec.

pub fn try_reserve(&mut self, additional: usize) -> Result<(), TryReserveError>

Invokes try_reserve on the underlying instance of Vec.

pub fn reserve_exact(&mut self, additional: usize)

Invokes reserve_exact on the underlying instance of Vec.

pub fn try_reserve_exact( &mut self, additional: usize, ) -> Result<(), TryReserveError>

Invokes try_reserve_exact on the underlying instance of Vec.

pub fn shrink_to_fit(&mut self)

Invokes shrink_to_fit on the underlying instance of Vec.

pub fn shrink_to(&mut self, min_capacity: usize)

Invokes shrink_to on the underlying instance of Vec.