typed_path

Struct Utf8Path

pub struct Utf8Path<T>
where
    T: Utf8Encoding,
{ /* private fields */ }

A slice of a path (akin to str).

This type supports a number of operations for inspecting a path, including breaking the path into its components (separated by / on Unix and by either / or \ on Windows), extracting the file name, determining whether the path is absolute, and so on.

This is an unsized type, meaning that it must always be used behind a pointer like & or Box. For an owned version of this type, see Utf8PathBuf.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding,
//       but all encodings work on all operating systems, providing the
//       ability to parse and operate on paths independently of the
//       compiled platform
let path = Utf8Path::<Utf8UnixEncoding>::new("./foo/bar.txt");

let parent = path.parent();
assert_eq!(parent, Some(Utf8Path::new("./foo")));

let file_stem = path.file_stem();
assert_eq!(file_stem, Some("bar"));

let extension = path.extension();
assert_eq!(extension, Some("txt"));

In addition to explicitly using Utf8Encodings, you can also leverage aliases available from the crate to work with paths:

use typed_path::{Utf8UnixPath, Utf8WindowsPath};

// Same as Utf8Path<Utf8UnixEncoding>
let path = Utf8UnixPath::new("/foo/bar.txt");

// Same as Utf8Path<Utf8WindowsEncoding>
let path = Utf8WindowsPath::new(r"C:\foo\bar.txt");

To mirror the design of Rust’s standard library, you can access the path associated with the compiled rust platform using Utf8NativePath, which itself is an alias to one of the other choices:

use typed_path::Utf8NativePath;

// On Unix, this would be Utf8UnixPath aka Utf8Path<Utf8UnixEncoding>
// On Windows, this would be Utf8WindowsPath aka Utf8Path<Utf8WindowsEncoding>
let path = Utf8NativePath::new("/foo/bar.txt");

Implementations

impl<T> Utf8Path<T>
where T: Utf8Encoding,

pub fn new<S: AsRef<str> + ?Sized>(s: &S) -> &Self

Directly wraps a str slice as a Utf8Path slice.

This is a cost-free conversion.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
Utf8Path::<Utf8UnixEncoding>::new("foo.txt");

You can create Utf8Paths from Strings, or even other Utf8Paths:

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let string = String::from("foo.txt");
let from_string = Utf8Path::<Utf8UnixEncoding>::new(&string);
let from_path = Utf8Path::new(&from_string);
assert_eq!(from_string, from_path);

There are also handy aliases to the Utf8Path with Utf8Encoding:

use typed_path::Utf8UnixPath;

let string = String::from("foo.txt");
let from_string = Utf8UnixPath::new(&string);
let from_path = Utf8UnixPath::new(&from_string);
assert_eq!(from_string, from_path);

pub fn as_str(&self) -> &str

Yields the underlying str slice.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let s = Utf8Path::<Utf8UnixEncoding>::new("foo.txt").as_str();
assert_eq!(s, "foo.txt");

pub fn to_path_buf(&self) -> Utf8PathBuf<T>

Converts a Utf8Path to an owned Utf8PathBuf.

Examples

use typed_path::{Utf8Path, Utf8PathBuf, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let path_buf = Utf8Path::<Utf8UnixEncoding>::new("foo.txt").to_path_buf();
assert_eq!(path_buf, Utf8PathBuf::from("foo.txt"));

pub fn is_absolute(&self) -> bool

Returns true if the Utf8Path is absolute, i.e., if it is independent of the current directory.

On Unix (Utf8UnixPath]), a path is absolute if it starts with the root, so is_absolute and has_root are equivalent.

On Windows (Utf8WindowsPath), a path is absolute if it has a prefix and starts with the root: c:\windows is absolute, while c:temp and \temp are not.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
assert!(!Utf8Path::<Utf8UnixEncoding>::new("foo.txt").is_absolute());

pub fn is_relative(&self) -> bool

Returns true if the Utf8Path is relative, i.e., not absolute.

See is_absolute’s documentation for more details.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
assert!(Utf8Path::<Utf8UnixEncoding>::new("foo.txt").is_relative());

pub fn is_valid(&self) -> bool

Returns true if the path is valid, meaning that all of its components are valid.

See Utf8Component::is_valid’s documentation for more details.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
assert!(Utf8Path::<Utf8UnixEncoding>::new("foo.txt").is_valid());
assert!(!Utf8Path::<Utf8UnixEncoding>::new("foo\0.txt").is_valid());

pub fn has_root(&self) -> bool

Returns true if the Utf8Path has a root.

On Unix (Utf8UnixPath), a path has a root if it begins with /.

On Windows (Utf8WindowsPath), a path has a root if it:

has no prefix and begins with a separator, e.g., \windows

has a prefix followed by a separator, e.g., c:\windows but not c:windows

has any non-disk prefix, e.g., \\server\share

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
assert!(Utf8Path::<Utf8UnixEncoding>::new("/etc/passwd").has_root());

pub fn parent(&self) -> Option<&Self>

Returns the Utf8Path without its final component, if there is one.

Returns None if the path terminates in a root or prefix.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let path = Utf8Path::<Utf8UnixEncoding>::new("/foo/bar");
let parent = path.parent().unwrap();
assert_eq!(parent, Utf8Path::new("/foo"));

let grand_parent = parent.parent().unwrap();
assert_eq!(grand_parent, Utf8Path::new("/"));
assert_eq!(grand_parent.parent(), None);

pub fn ancestors(&self) -> Utf8Ancestors<'_, T>

Produces an iterator over Utf8Path and its ancestors.

The iterator will yield the Utf8Path that is returned if the parent method is used zero or more times. That means, the iterator will yield &self, &self.parent().unwrap(), &self.parent().unwrap().parent().unwrap() and so on. If the parent method returns None, the iterator will do likewise. The iterator will always yield at least one value, namely &self.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let mut ancestors = Utf8Path::<Utf8UnixEncoding>::new("/foo/bar").ancestors();
assert_eq!(ancestors.next(), Some(Utf8Path::new("/foo/bar")));
assert_eq!(ancestors.next(), Some(Utf8Path::new("/foo")));
assert_eq!(ancestors.next(), Some(Utf8Path::new("/")));
assert_eq!(ancestors.next(), None);

// NOTE: A path cannot be created on its own without a defined encoding
let mut ancestors = Utf8Path::<Utf8UnixEncoding>::new("../foo/bar").ancestors();
assert_eq!(ancestors.next(), Some(Utf8Path::new("../foo/bar")));
assert_eq!(ancestors.next(), Some(Utf8Path::new("../foo")));
assert_eq!(ancestors.next(), Some(Utf8Path::new("..")));
assert_eq!(ancestors.next(), Some(Utf8Path::new("")));
assert_eq!(ancestors.next(), None);

pub fn file_name(&self) -> Option<&str>

Returns the final component of the Utf8Path, if there is one.

If the path is a normal file, this is the file name. If it’s the path of a directory, this is the directory name.

Returns None if the path terminates in ...

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
assert_eq!(Some("bin"), Utf8Path::<Utf8UnixEncoding>::new("/usr/bin/").file_name());
assert_eq!(Some("foo.txt"), Utf8Path::<Utf8UnixEncoding>::new("tmp/foo.txt").file_name());
assert_eq!(Some("foo.txt"), Utf8Path::<Utf8UnixEncoding>::new("foo.txt/.").file_name());
assert_eq!(Some("foo.txt"), Utf8Path::<Utf8UnixEncoding>::new("foo.txt/.//").file_name());
assert_eq!(None, Utf8Path::<Utf8UnixEncoding>::new("foo.txt/..").file_name());
assert_eq!(None, Utf8Path::<Utf8UnixEncoding>::new("/").file_name());

pub fn strip_prefix<P>(&self, base: P) -> Result<&Utf8Path<T>, StripPrefixError>
where P: AsRef<Utf8Path<T>>,

Returns a path that, when joined onto base, yields self.

Errors

If base is not a prefix of self (i.e., starts_with returns false), returns Err.

Examples

use typed_path::{Utf8Path, Utf8PathBuf, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let path = Utf8Path::<Utf8UnixEncoding>::new("/test/haha/foo.txt");

assert_eq!(path.strip_prefix("/"), Ok(Utf8Path::new("test/haha/foo.txt")));
assert_eq!(path.strip_prefix("/test"), Ok(Utf8Path::new("haha/foo.txt")));
assert_eq!(path.strip_prefix("/test/"), Ok(Utf8Path::new("haha/foo.txt")));
assert_eq!(path.strip_prefix("/test/haha/foo.txt"), Ok(Utf8Path::new("")));
assert_eq!(path.strip_prefix("/test/haha/foo.txt/"), Ok(Utf8Path::new("")));

assert!(path.strip_prefix("test").is_err());
assert!(path.strip_prefix("/haha").is_err());

let prefix = Utf8PathBuf::<Utf8UnixEncoding>::from("/test/");
assert_eq!(path.strip_prefix(prefix), Ok(Utf8Path::new("haha/foo.txt")));

pub fn starts_with<P>(&self, base: P) -> bool
where P: AsRef<Utf8Path<T>>,

Determines whether base is a prefix of self.

Only considers whole path components to match.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let path = Utf8Path::<Utf8UnixEncoding>::new("/etc/passwd");

assert!(path.starts_with("/etc"));
assert!(path.starts_with("/etc/"));
assert!(path.starts_with("/etc/passwd"));
assert!(path.starts_with("/etc/passwd/")); // extra slash is okay
assert!(path.starts_with("/etc/passwd///")); // multiple extra slashes are okay

assert!(!path.starts_with("/e"));
assert!(!path.starts_with("/etc/passwd.txt"));

assert!(!Utf8Path::<Utf8UnixEncoding>::new("/etc/foo.rs").starts_with("/etc/foo"));

pub fn ends_with<P>(&self, child: P) -> bool
where P: AsRef<Utf8Path<T>>,

Determines whether child is a suffix of self.

Only considers whole path components to match.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let path = Utf8Path::<Utf8UnixEncoding>::new("/etc/resolv.conf");

assert!(path.ends_with("resolv.conf"));
assert!(path.ends_with("etc/resolv.conf"));
assert!(path.ends_with("/etc/resolv.conf"));

assert!(!path.ends_with("/resolv.conf"));
assert!(!path.ends_with("conf")); // use .extension() instead

pub fn file_stem(&self) -> Option<&str>

Extracts the stem (non-extension) portion of self.file_name.

The stem is:

None, if there is no file name;

The entire file name if there is no embedded .;

The entire file name if the file name begins with . and has no other .s within;

Otherwise, the portion of the file name before the final .

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
assert_eq!("foo", Utf8Path::<Utf8UnixEncoding>::new("foo.rs").file_stem().unwrap());
assert_eq!("foo.tar", Utf8Path::<Utf8UnixEncoding>::new("foo.tar.gz").file_stem().unwrap());

pub fn extension(&self) -> Option<&str>

Extracts the extension of self.file_name, if possible.

The extension is:

None, if there is no file name;

None, if there is no embedded .;

None, if the file name begins with . and has no other .s within;

Otherwise, the portion of the file name after the final .

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
assert_eq!("rs", Utf8Path::<Utf8UnixEncoding>::new("foo.rs").extension().unwrap());
assert_eq!("gz", Utf8Path::<Utf8UnixEncoding>::new("foo.tar.gz").extension().unwrap());

pub fn normalize(&self) -> Utf8PathBuf<T>

Returns an owned Utf8PathBuf by resolving .. and . segments.

When multiple, sequential path segment separation characters are found (e.g. / for Unix and either \ or / on Windows), they are replaced by a single instance of the platform-specific path segment separator (/ on Unix and \ on Windows).

Examples

use typed_path::{Utf8Path, Utf8PathBuf, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
assert_eq!(
    Utf8Path::<Utf8UnixEncoding>::new("foo/bar//baz/./asdf/quux/..").normalize(),
    Utf8PathBuf::from("foo/bar/baz/asdf"),
);

When starting with a root directory, any .. segment whose parent is the root directory will be filtered out:

use typed_path::{Utf8Path, Utf8PathBuf, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
assert_eq!(
    Utf8Path::<Utf8UnixEncoding>::new("/../foo").normalize(),
    Utf8PathBuf::from("/foo"),
);

If any .. is left unresolved as the path is relative and no parent is found, it is discarded:

use typed_path::{Utf8Path, Utf8PathBuf, Utf8UnixEncoding, Utf8WindowsEncoding};

assert_eq!(
    Utf8Path::<Utf8UnixEncoding>::new("../foo/..").normalize(),
    Utf8PathBuf::from(""),
);

//Windows prefixes also count this way, but the prefix remains
assert_eq!(
    Utf8Path::<Utf8WindowsEncoding>::new(r"C:..\foo\..").normalize(),
    Utf8PathBuf::from(r"C:"),
);

pub fn absolutize(&self) -> Result<Utf8PathBuf<T>>

Converts a path to an absolute form by normalizing the path, returning a Utf8PathBuf.

In the case that the path is relative, the current working directory is prepended prior to normalizing.

Examples

use typed_path::{utils, Utf8Path, Utf8UnixEncoding};

// With an absolute path, it is just normalized
let path = Utf8Path::<Utf8UnixEncoding>::new("/a/b/../c/./d");
assert_eq!(path.absolutize().unwrap(), Utf8Path::new("/a/c/d"));

// With a relative path, it is first joined with the current working directory
// and then normalized
let cwd = utils::utf8_current_dir().unwrap().with_encoding::<Utf8UnixEncoding>();
let path = cwd.join(Utf8Path::new("a/b/../c/./d"));
assert_eq!(path.absolutize().unwrap(), cwd.join(Utf8Path::new("a/c/d")));

pub fn join<P: AsRef<Utf8Path<T>>>(&self, path: P) -> Utf8PathBuf<T>

Creates an owned Utf8PathBuf with path adjoined to self.

See Utf8PathBuf::push for more details on what it means to adjoin a path.

Examples

use typed_path::{Utf8Path, Utf8PathBuf, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
assert_eq!(
    Utf8Path::<Utf8UnixEncoding>::new("/etc").join("passwd"),
    Utf8PathBuf::from("/etc/passwd"),
);

pub fn join_checked<P: AsRef<Utf8Path<T>>>( &self, path: P, ) -> Result<Utf8PathBuf<T>, CheckedPathError>

Creates an owned Utf8PathBuf with path adjoined to self, checking the path to ensure it is safe to join. When dealing with user-provided paths, this is the preferred method.

See Utf8PathBuf::push_checked for more details on what it means to adjoin a path safely.

Examples

use typed_path::{CheckedPathError, Utf8Path, Utf8PathBuf, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let path = Utf8Path::<Utf8UnixEncoding>::new("/etc");

// A valid path can be joined onto the existing one
assert_eq!(path.join_checked("passwd"), Ok(Utf8PathBuf::from("/etc/passwd")));

// An invalid path will result in an error
assert_eq!(path.join_checked("/sneaky/replacement"), Err(CheckedPathError::UnexpectedRoot));

pub fn with_file_name<S: AsRef<str>>(&self, file_name: S) -> Utf8PathBuf<T>

Creates an owned Utf8PathBuf like self but with the given file name.

See Utf8PathBuf::set_file_name for more details.

Examples

use typed_path::{Utf8Path, Utf8PathBuf, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let path = Utf8Path::<Utf8UnixEncoding>::new("/tmp/foo.txt");
assert_eq!(path.with_file_name("bar.txt"), Utf8PathBuf::from("/tmp/bar.txt"));

// NOTE: A path cannot be created on its own without a defined encoding
let path = Utf8Path::<Utf8UnixEncoding>::new("/tmp");
assert_eq!(path.with_file_name("var"), Utf8PathBuf::from("/var"));

pub fn with_extension<S: AsRef<str>>(&self, extension: S) -> Utf8PathBuf<T>

Creates an owned Utf8PathBuf like self but with the given extension.

See Utf8PathBuf::set_extension for more details.

Examples

use typed_path::{Utf8Path, Utf8PathBuf, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let path = Utf8Path::<Utf8UnixEncoding>::new("foo.rs");
assert_eq!(path.with_extension("txt"), Utf8PathBuf::from("foo.txt"));

// NOTE: A path cannot be created on its own without a defined encoding
let path = Utf8Path::<Utf8UnixEncoding>::new("foo.tar.gz");
assert_eq!(path.with_extension(""), Utf8PathBuf::from("foo.tar"));
assert_eq!(path.with_extension("xz"), Utf8PathBuf::from("foo.tar.xz"));
assert_eq!(path.with_extension("").with_extension("txt"), Utf8PathBuf::from("foo.txt"));

pub fn components(&self) -> <T as Utf8Encoding>::Components<'_>

Produces an iterator over the Utf8Components of the path.

When parsing the path, there is a small amount of normalization:

Repeated separators are ignored, so a/b and a//b both have a and b as components.

Occurrences of . are normalized away, except if they are at the beginning of the path. For example, a/./b, a/b/, a/b/. and a/b all have a and b as components, but ./a/b starts with an additional CurDir component.

A trailing slash is normalized away, /a/b and /a/b/ are equivalent.

Note that no other normalization takes place; in particular, a/c and a/b/../c are distinct, to account for the possibility that b is a symbolic link (so its parent isn’t a).

Examples

use typed_path::{Utf8Path, Utf8UnixComponent, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let mut components = Utf8Path::<Utf8UnixEncoding>::new("/tmp/foo.txt").components();

assert_eq!(components.next(), Some(Utf8UnixComponent::RootDir));
assert_eq!(components.next(), Some(Utf8UnixComponent::Normal("tmp")));
assert_eq!(components.next(), Some(Utf8UnixComponent::Normal("foo.txt")));
assert_eq!(components.next(), None)

pub fn iter(&self) -> Utf8Iter<'_, T>

Produces an iterator over the path’s components viewed as str slices.

For more information about the particulars of how the path is separated into components, see components.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding};

// NOTE: A path cannot be created on its own without a defined encoding
let mut it = Utf8Path::<Utf8UnixEncoding>::new("/tmp/foo.txt").iter();

assert_eq!(it.next(), Some(typed_path::constants::unix::SEPARATOR_STR));
assert_eq!(it.next(), Some("tmp"));
assert_eq!(it.next(), Some("foo.txt"));
assert_eq!(it.next(), None)

pub fn with_encoding<U>(&self) -> Utf8PathBuf<U>
where U: Utf8Encoding,

Creates an owned Utf8PathBuf like self but with a different encoding.

Note

As part of the process of converting between encodings, the path will need to be rebuilt. This involves pushing each component, which may result in differences in the resulting path such as resolving . and .. early or other unexpected side effects.

Examples

use typed_path::{Utf8Path, Utf8UnixEncoding, Utf8WindowsEncoding};

// Convert from Unix to Windows
let unix_path = Utf8Path::<Utf8UnixEncoding>::new("/tmp/foo.txt");
let windows_path = unix_path.with_encoding::<Utf8WindowsEncoding>();
assert_eq!(windows_path, Utf8Path::<Utf8WindowsEncoding>::new(r"\tmp\foo.txt"));

// Converting from Windows to Unix will drop any prefix
let windows_path = Utf8Path::<Utf8WindowsEncoding>::new(r"C:\tmp\foo.txt");
let unix_path = windows_path.with_encoding::<Utf8UnixEncoding>();
assert_eq!(unix_path, Utf8Path::<Utf8UnixEncoding>::new(r"/tmp/foo.txt"));

// Converting to itself should retain everything
let path = Utf8Path::<Utf8WindowsEncoding>::new(r"C:\tmp\foo.txt");
assert_eq!(
    path.with_encoding::<Utf8WindowsEncoding>(),
    Utf8Path::<Utf8WindowsEncoding>::new(r"C:\tmp\foo.txt"),
);

pub fn with_encoding_checked<U>( &self, ) -> Result<Utf8PathBuf<U>, CheckedPathError>
where U: Utf8Encoding,

Like with_encoding, creates an owned Utf8PathBuf like self but with a different encoding. Additionally, checks to ensure that the produced path will be valid.

Note

As part of the process of converting between encodings, the path will need to be rebuilt. This involves pushing and checking each component, which may result in differences in the resulting path such as resolving . and .. early or other unexpected side effects.

Examples

use typed_path::{CheckedPathError, Utf8Path, Utf8UnixEncoding, Utf8WindowsEncoding};

// Convert from Unix to Windows
let unix_path = Utf8Path::<Utf8UnixEncoding>::new("/tmp/foo.txt");
let windows_path = unix_path.with_encoding_checked::<Utf8WindowsEncoding>().unwrap();
assert_eq!(windows_path, Utf8Path::<Utf8WindowsEncoding>::new(r"\tmp\foo.txt"));

// Converting from Windows to Unix will drop any prefix
let windows_path = Utf8Path::<Utf8WindowsEncoding>::new(r"C:\tmp\foo.txt");
let unix_path = windows_path.with_encoding_checked::<Utf8UnixEncoding>().unwrap();
assert_eq!(unix_path, Utf8Path::<Utf8UnixEncoding>::new(r"/tmp/foo.txt"));

// Converting from Unix to Windows with invalid filename characters like `|` should fail
let unix_path = Utf8Path::<Utf8UnixEncoding>::new("/|invalid|/foo.txt");
assert_eq!(
    unix_path.with_encoding_checked::<Utf8WindowsEncoding>(),
    Err(CheckedPathError::InvalidFilename),
);

// Converting from Unix to Windows with unexpected prefix embedded in path should fail
let unix_path = Utf8Path::<Utf8UnixEncoding>::new("/path/c:/foo.txt");
assert_eq!(
    unix_path.with_encoding_checked::<Utf8WindowsEncoding>(),
    Err(CheckedPathError::UnexpectedPrefix),
);

pub fn into_path_buf(self: Box<Utf8Path<T>>) -> Utf8PathBuf<T>

Converts a Box<Utf8Path> into a Utf8PathBuf without copying or allocating.

pub fn from_bytes_path<U>(path: &Path<U>) -> Result<&Self, Utf8Error>
where U: Encoding,

Converts a non-UTF-8 Path to a UTF-8 Utf8PathBuf by checking that the path contains valid UTF-8.

Errors

Returns Err if the path is not UTF-8 with a description as to why the provided component is not UTF-8.

Examples

use typed_path::{Path, Utf8Path, UnixEncoding, Utf8UnixEncoding};

let path = Path::<UnixEncoding>::new(&[0xf0, 0x9f, 0x92, 0x96]);
let utf8_path = Utf8Path::<Utf8UnixEncoding>::from_bytes_path(&path).unwrap();
assert_eq!(utf8_path.as_str(), "💖");

pub unsafe fn from_bytes_path_unchecked<U>(path: &Path<U>) -> &Self
where U: Encoding,

Converts a non-UTF-8 Path to a UTF-8 Utf8Path without checking that the path contains valid UTF-8.

See the safe version, from_bytes_path, for more information.

Safety

The path passed in must be valid UTF-8.

Examples

use typed_path::{Path, Utf8Path, UnixEncoding, Utf8UnixEncoding};

let path = Path::<UnixEncoding>::new(&[0xf0, 0x9f, 0x92, 0x96]);
let utf8_path = unsafe {
    Utf8Path::<Utf8UnixEncoding>::from_bytes_path_unchecked(&path)
};
assert_eq!(utf8_path.as_str(), "💖");

pub fn as_bytes_path<U>(&self) -> &Path<U>
where U: Encoding,

Converts a UTF-8 Utf8Path to a non-UTF-8 Path.

Examples

use typed_path::{Path, Utf8Path, UnixEncoding, Utf8UnixEncoding};

let utf8_path = Utf8Path::<Utf8UnixEncoding>::new("💖");
let path = utf8_path.as_bytes_path::<UnixEncoding>();
assert_eq!(path.as_bytes(), &[0xf0, 0x9f, 0x92, 0x96]);

impl<T> Utf8Path<T>
where T: Utf8Encoding,

pub fn has_platform_encoding(&self) -> bool
where T: 'static,

Returns true if the encoding is the platform abstraction (Utf8PlatformEncoding), otherwise returns false.

Examples

use typed_path::{Utf8PlatformPath, Utf8UnixPath, Utf8WindowsPath};

assert!(Utf8PlatformPath::new("/some/path").has_platform_encoding());
assert!(!Utf8UnixPath::new("/some/path").has_platform_encoding());
assert!(!Utf8WindowsPath::new("/some/path").has_platform_encoding());

pub fn with_platform_encoding(&self) -> Utf8PathBuf<Utf8PlatformEncoding>

Creates an owned Utf8PathBuf like self but using Utf8PlatformEncoding.

See Utf8Path::with_encoding for more information.

pub fn with_platform_encoding_checked( &self, ) -> Result<Utf8PathBuf<Utf8PlatformEncoding>, CheckedPathError>

Creates an owned Utf8PathBuf like self but using Utf8PlatformEncoding, ensuring it is a valid platform path.

See Utf8Path::with_encoding_checked for more information.

impl<T> Utf8Path<T>
where T: Utf8Encoding,

pub fn has_unix_encoding(&self) -> bool

Returns true if the encoding for the path is for Unix.

Examples

use typed_path::{Utf8UnixPath, Utf8WindowsPath};

assert!(Utf8UnixPath::new("/some/path").has_unix_encoding());
assert!(!Utf8WindowsPath::new(r"\some\path").has_unix_encoding());

pub fn with_unix_encoding(&self) -> Utf8PathBuf<Utf8UnixEncoding>

Creates an owned Utf8PathBuf like self but using Utf8UnixEncoding.

See Utf8Path::with_encoding for more information.

pub fn with_unix_encoding_checked( &self, ) -> Result<Utf8PathBuf<Utf8UnixEncoding>, CheckedPathError>

Creates an owned Utf8PathBuf like self but using Utf8UnixEncoding, ensuring it is a valid Unix path.

See Utf8Path::with_encoding_checked for more information.

impl Utf8Path<Utf8UnixEncoding>

pub fn to_typed_path(&self) -> Utf8TypedPath<'_>

pub fn to_typed_path_buf(&self) -> Utf8TypedPathBuf

impl<T> Utf8Path<T>
where T: Utf8Encoding,

pub fn has_windows_encoding(&self) -> bool

Returns true if the encoding for the path is for Windows.

Examples

use typed_path::{Utf8UnixPath, Utf8WindowsPath};

assert!(!Utf8UnixPath::new("/some/path").has_windows_encoding());
assert!(Utf8WindowsPath::new(r"\some\path").has_windows_encoding());

pub fn with_windows_encoding(&self) -> Utf8PathBuf<Utf8WindowsEncoding>

Creates an owned Utf8PathBuf like self but using Utf8WindowsEncoding.

See Utf8Path::with_encoding for more information.

pub fn with_windows_encoding_checked( &self, ) -> Result<Utf8PathBuf<Utf8WindowsEncoding>, CheckedPathError>

Creates an owned Utf8PathBuf like self but using Utf8WindowsEncoding, ensuring it is a valid Windows path.

See Utf8Path::with_encoding_checked for more information.

impl Utf8Path<Utf8WindowsEncoding>

pub fn to_typed_path(&self) -> Utf8TypedPath<'_>

pub fn to_typed_path_buf(&self) -> Utf8TypedPathBuf